Added: Enemy, Health Logic, Mech States and FSM
- Enemy:
Enemy is in the dev room, and has health. Renders perfectly.
- Health Logic:
There's a separate health script with values for max and current health. Also has functions for adding current and max health, and taking away current health
Mech FSM:
Made a state machine for the mech, and states for idle, inactive, getting and losing a pilot. States are barebones

**Added a Todo list