Separated FOV between on-foot and in-mech, gave player ability to shoot, gave character ability to get hurt
There are now 2 diff camera settings, a 'player cam' for on foot action and a 'mech cam' for being in the mech. The art set when entering the mech state and leaving the mech state

There's a script that handles player character attacking.

Made Damageable interface, utilized Health script for enemies